+++ quasar-utils/src/main/groovy/net/corda/plugins/QuasarExtension.groovy
@@ -1,5 +1,6 @@
 package net.corda.plugins
 
+import groovy.transform.PackageScope
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
@@ -9,16 +10,59 @@ import javax.inject.Inject
 
 class QuasarExtension {
 
+    /**
+     * Maven group for the Quasar agent.
+     */
+    final Property<String> group
+
+    /**
+     * Maven version for the Quasar agent.
+     */
+    final Property<String> version
+
+    /**
+     * Maven classifier for the Quasar agent.
+     */
+    final Property<String> classifier
+
+    /**
+     * Dependency notation for the Quasar agent to use.
+     */
+    final Provider<Map<String, String>> dependency
+
+    /**
+     * Runtime options for the Quasar agent:
+     * - debug
+     * - verbose
+     * - globs of packages not to instrument.
+     */
     final Property<Boolean> debug
 
     final Property<Boolean> verbose
 
     final ListProperty<String> excludePackages
 
+    @PackageScope
     final Provider<String> options
 
     @Inject
-    QuasarExtension(ObjectFactory objects, Iterable<? extends String> initialExclusions) {
+    QuasarExtension(
+        ObjectFactory objects,
+        String defaultGroup,
+        String defaultVersion,
+        String defaultClassifier,
+        Iterable<? extends String> initialExclusions
+    ) {
+        group = objects.property(String)
+        group.set(defaultGroup)
+        version = objects.property(String)
+        version.set(defaultVersion)
+        classifier = objects.property(String)
+        classifier.set(defaultClassifier)
+        dependency = group.map { grp ->
+            [ group: grp, name: 'quasar-core', version: version.get(), classifier: classifier.get(), ext: 'jar' ]
+        }
+
         debug = objects.property(Boolean)
         debug.set(false)
         verbose = objects.property(Boolean)